digital_marketing/requirements.txt
django
psycopg2-binary
python-dotenv
djangorestframework
django-cors-headers
PyJWT
requests
facebook-business
google-auth-oauthlib
google-auth
dj-database-url
python-decouple

